Fixed most gameplay issues.
Drone movement is fixed. HUD works VIA shipping system. Containers can be picked up, Moved, lowered/ delivered.

Need to build out the level, Need to finish Container health system/ death function, Need to test and fix the Scoring system, Need to fix the Cargo score in HUD (Should count down as points are lost).

+++ Source/ZeroGDelivery/Public/HUD/GameHUD.h
@@ -29,10 +29,10 @@ public:
 	void UpdateScore(int32 Score);
 
 	UFUNCTION()
-	void UpdateTimer(float TimeRemaining);
+	void UpdateTimer(int32 TimeRemaining);
 
 	UFUNCTION()
-	void UpdateContainerHealth(int32 HealthPercent);
+	void UpdateContainerHealth(float HealthPercent);
 
 	UFUNCTION()
 	void AddCargoScore(int FinalCargoScore);
